+++ src/pages/UpdateGroup/UpdateGroup.module.css
@@ -1,0 +1,5 @@
+.container {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+}
